feat: #WB2-1530, add edit content page
